renaming LinkAndStreamSettings header file to ModuleSettings
