+++ src/CMakeLists.txt
@@ -175,7 +175,7 @@ set(kafka_to_nexus_INC
         TimeUtility.h
         HDFOperations.h
         HDFVersionCheck.h
-        LinkAndStreamSettings.h
+        ModuleSettings.h
         StreamHDFInfo.h
         CommandSystem/CommandListener.h
         CommandSystem/FeedbackProducerBase.h
